Fixes CollectionView to set the correct Delegate
