+++ main/src/addins/MonoDevelop.DesignerSupport/MonoDevelop.DesignerSupport.csproj
@@ -133,7 +133,7 @@
     <Compile Include="MonoDevelop.DesignerSupport.Toolbox\ProgressIndicator.cs" />
     <Compile Include="MonoDevelop.DesignerSupport.Toolbox\CollectionView\CollectionView.cs" />
     <Compile Include="MonoDevelop.DesignerSupport.Toolbox\CollectionView\CollectionViewDataSource.cs" />
-    <Compile Include="MonoDevelop.DesignerSupport.Toolbox\CollectionView\CollectionViewFlowLayout.cs" />
+    <Compile Include="MonoDevelop.DesignerSupport.Toolbox\CollectionView\CollectionViewFlowLayoutDelegate.cs" />
     <Compile Include="MonoDevelop.DesignerSupport.Toolbox\CollectionView\CollectionViewItems.cs" />
     <Compile Include="MonoDevelop.DesignerSupport.Toolbox\IToolboxWidget.cs" />
     <Compile Include="MonoDevelop.DesignerSupport.Toolbox\NativeViews\SearchTextField.cs" />
@@ -144,7 +144,6 @@
     <Compile Include="MonoDevelop.DesignerSupport.Toolbox\CollectionView\Styles.cs" />
     <Compile Include="MonoDevelop.DesignerSupport.Toolbox\NativeViews\PaddingView.cs" />
     <Compile Include="MonoDevelop.DesignerSupport.Toolbox\NativeViews\ExpanderButton.cs" />
-    <Compile Include="MonoDevelop.DesignerSupport.Toolbox\CollectionView\ColleccionViewDelegate.cs" />
   </ItemGroup>
   <ItemGroup>
     <EmbeddedResource Include="MonoDevelop.DesignerSupport.addin.xml" />
